refactor(server): extract email-related routes to a separate module
