+++ lib/routes/account.js
@@ -4,7 +4,6 @@
 
 'use strict'
 
-const butil = require('../crypto/butil')
 const error = require('../error')
 const isA = require('joi')
 const METRICS_CONTEXT_SCHEMA = require('../metrics/context').schema
@@ -26,18 +25,7 @@ const MS_ONE_DAY = 1000 * 60 * 60 * 24
 const MS_ONE_WEEK = MS_ONE_DAY * 7
 const MS_ONE_MONTH = MS_ONE_DAY * 30
 
-module.exports = (
-  log,
-  db,
-  bounces,
-  mailer,
-  Password,
-  config,
-  customs,
-  checkPassword,
-  push
-  ) => {
-
+module.exports = (log, db, mailer, Password, config, customs, checkPassword, push) => {
   const getGeoData = require('../geodb')(log)
   const features = require('../features')(config)
   const verificationReminder = require('../verification-reminders')(log, db)
@@ -1185,677 +1173,6 @@ module.exports = (
             }
           )
           .then(reply, reply)
-      }
-    },
-    {
-      method: 'GET',
-      path: '/recovery_email/check_can_add_secondary_address',
-      config: {
-        auth: {
-          strategy: 'sessionTokenWithVerificationStatus'
-        }
-      },
-      handler (request, reply) {
-        log.begin('Account.RecoveryEmailEnabled', request)
-        const sessionToken = request.auth.credentials
-        let enabled = false
-
-        // Secondary emails are enabled if feature is enabled for user and they are in a verified session
-        if (features.isSecondaryEmailEnabled(sessionToken.email) && sessionToken.tokenVerified) {
-          enabled = true
-        }
-
-        return reply({ok: enabled})
-      }
-    },
-    {
-      method: 'GET',
-      path: '/recovery_email/status',
-      config: {
-        auth: {
-          strategy: 'sessionTokenWithVerificationStatus'
-        },
-        validate: {
-          query: {
-            reason: isA.string().max(16).optional()
-          }
-        },
-        response: {
-          schema: {
-            // There's code in the handler that checks for a valid email,
-            // no point adding overhead by doing it again here.
-            email: isA.string().required(),
-            verified: isA.boolean().required(),
-            sessionVerified: isA.boolean().optional(),
-            emailVerified: isA.boolean().optional()
-          }
-        }
-      },
-      handler: function (request, reply) {
-        log.begin('Account.RecoveryEmailStatus', request)
-        var sessionToken = request.auth.credentials
-        if (request.query && request.query.reason === 'push') {
-          // log to the push namespace that account was verified via push
-          log.info({
-            op: 'push.pushToDevices',
-            name: 'recovery_email_reason.push'
-          })
-        }
-
-        cleanUpIfAccountInvalid()
-          .then(createResponse)
-          .then(reply, reply)
-
-        function cleanUpIfAccountInvalid() {
-          // Some historical bugs mean we've allowed creation
-          // of accounts with invalid email addresses. These
-          // can never be verified, so the best we can do is
-          // to delete them so the browser will stop polling.
-          if (! sessionToken.emailVerified) {
-            if (! validators.isValidEmailAddress(sessionToken.email)) {
-              return db.deleteAccount(sessionToken)
-                .then(
-                  function () {
-                    // Act as though we deleted the account asynchronously
-                    // and caused the sessionToken to become invalid.
-                    throw error.invalidToken('This account was invalid and has been deleted')
-                  }
-                )
-            }
-          }
-          return P.resolve()
-        }
-
-        function createResponse() {
-
-          var sessionVerified = sessionToken.tokenVerified
-          var emailVerified = !! sessionToken.emailVerified
-
-          // For backwards-compatibility reasons, the reported verification status
-          // depends on whether the sessionToken was created with keys=true and
-          // whether it has subsequently been verified.  If it was created with
-          // keys=true then we musn't say verified=true until the session itself
-          // has been verified.  Otherwise, desktop clients will attempt to use
-          // an unverified session to connect to sync, and produce a very confusing
-          // user experience.
-          var isVerified = emailVerified
-          if (sessionToken.mustVerify) {
-            isVerified = isVerified && sessionVerified
-          }
-
-          return {
-            email: sessionToken.email,
-            verified: isVerified,
-            sessionVerified: sessionVerified,
-            emailVerified: emailVerified
-          }
-        }
-      }
-    },
-    {
-      method: 'POST',
-      path: '/recovery_email/resend_code',
-      config: {
-        auth: {
-          strategy: 'sessionTokenWithVerificationStatus'
-        },
-        validate: {
-          query: {
-            service: validators.service
-          },
-          payload: {
-            email: validators.email().optional(),
-            service: validators.service,
-            redirectTo: validators.redirectTo(config.smtp.redirectDomain).optional(),
-            resume: isA.string().max(2048).optional(),
-            metricsContext: METRICS_CONTEXT_SCHEMA
-          }
-        }
-      },
-      handler: function (request, reply) {
-        log.begin('Account.RecoveryEmailResend', request)
-        const email = request.payload.email
-        const sessionToken = request.auth.credentials
-        const service = request.payload.service || request.query.service
-
-        // This endpoint can verify multiple types of codes, set these values once it
-        // is known what is being verified.
-        let code
-        let verifyFunction
-        let event
-        let emails = []
-
-        // Return immediately if this session or token is already verified. Only exception
-        // is if the email param has been specified, which means that this is
-        // a request to verify a secondary email.
-        if (sessionToken.emailVerified && sessionToken.tokenVerified && ! email) {
-          return reply({})
-        }
-
-        customs.check(request, sessionToken.email, 'recoveryEmailResendCode')
-          .then(setVerifyCode)
-          .then(setVerifyFunction)
-          .then(() => {
-            const mailerOpts = {
-              code: code,
-              service: service,
-              timestamp: Date.now(),
-              redirectTo: request.payload.redirectTo,
-              resume: request.payload.resume,
-              acceptLanguage: request.app.acceptLanguage,
-              uaBrowser: sessionToken.uaBrowser,
-              uaBrowserVersion: sessionToken.uaBrowserVersion,
-              uaOS: sessionToken.uaOS,
-              uaOSVersion: sessionToken.uaOSVersion,
-              uaDeviceType: sessionToken.uaDeviceType
-            }
-
-            return verifyFunction(emails, sessionToken, mailerOpts)
-          })
-          .then(() => request.emitMetricsEvent(`email.${event}.resent`))
-          .then(
-            () => reply({}),
-            reply
-          )
-
-        function setVerifyCode() {
-          return db.accountEmails(sessionToken.uid)
-            .then((emailData) => {
-              if (email) {
-                // If an email address is specified in payload, this is a request to verify
-                // a secondary email. This should return the corresponding email code for verification.
-                let emailVerified = false
-                emailData.some((userEmail) => {
-                  if (userEmail.normalizedEmail === email.toLowerCase()) {
-                    code = userEmail.emailCode
-                    emailVerified = userEmail.isVerified
-                    emails = [userEmail]
-                    return true
-                  }
-                })
-
-                // Don't resend code for already verified emails
-                if (emailVerified) {
-                  return reply({})
-                }
-              } else if (sessionToken.tokenVerificationId) {
-                emails = emailData
-                code = sessionToken.tokenVerificationId
-              } else {
-                code = sessionToken.emailCode
-              }
-            })
-        }
-
-        function setVerifyFunction() {
-          if (email) {
-            verifyFunction = mailer.sendVerifySecondaryEmail
-            event = 'verification_email'
-          } else if (! sessionToken.emailVerified) {
-            verifyFunction = mailer.sendVerifyCode
-            event = 'verification'
-          } else {
-            verifyFunction = mailer.sendVerifyLoginEmail
-            event = 'confirmation'
-          }
-        }
-      }
-    },
-    {
-      method: 'POST',
-      path: '/recovery_email/verify_code',
-      config: {
-        validate: {
-          query: {
-            service: validators.service,
-            reminder: isA.string().max(32).alphanum().optional(),
-            type: isA.string().max(32).alphanum().optional()
-          },
-          payload: {
-            uid: isA.string().max(32).regex(HEX_STRING).required(),
-            code: isA.string().min(32).max(32).regex(HEX_STRING).required(),
-            service: validators.service,
-            reminder: isA.string().max(32).alphanum().optional(),
-            type: isA.string().max(32).alphanum().optional(),
-            marketingOptIn: isA.boolean()
-          }
-        }
-      },
-      handler: function (request, reply) {
-        const uid = request.payload.uid
-        const code = request.payload.code
-        const service = request.payload.service || request.query.service
-        const reminder = request.payload.reminder || request.query.reminder
-        const type = request.payload.type || request.query.type
-        const marketingOptIn = request.payload.marketingOptIn
-
-        log.begin('Account.RecoveryEmailVerify', request)
-
-        // verify_code because we don't know what type this is yet, but
-        // we want to record right away before anything could fail, so
-        // we can see in a flow that a user tried to verify, even if it
-        // failed right away.
-        request.emitMetricsEvent('email.verify_code.clicked')
-
-        /**
-         * Below is a summary of the verify_code flow. This flow is used to verify emails, sign-in and
-         * account codes.
-         *
-         * 1) Check request against customs server, proceed if valid.
-         *
-         * 2) If type=`secondary` then this is an email code and verify it
-         *    accordingly.
-         *
-         * 3) Otherwise attempt to verify code as sign-in code then account code.
-         */
-        return db.account(uid)
-          .then((account) => {
-            // This endpoint is not authenticated, so we need to look up
-            // the target email address before we can check it with customs.
-            return customs.check(request, account.email, 'recoveryEmailVerifyCode')
-              .then(() => {
-                return account
-              })
-          })
-          .then((account) => {
-            // Check if param `type` is specified and equal to `secondary`
-            // If so, verify the secondary email and respond
-            if (type && type === 'secondary' && features.isSecondaryEmailEnabled(account.email)) {
-              let matchedEmail
-              return db.accountEmails(account.uid)
-                .then((emails) => {
-                  const isEmailVerification = emails.some((email) => {
-                    if (email.emailCode && (code === email.emailCode)) {
-                      matchedEmail = email
-                      log.info({
-                        op: 'account.verifyEmail.secondary.started',
-                        uid: uid,
-                        code: request.payload.code
-                      })
-                      return true
-                    }
-                  })
-
-                  // Attempt to verify email token not associated with account
-                  if (! isEmailVerification) {
-                    throw error.invalidVerificationCode()
-                  }
-
-                  // User is attempting to verify a secondary email that has already been verified.
-                  // Silently succeed and don't send post verification email.
-                  if (matchedEmail.isVerified) {
-                    log.info({
-                      op: 'account.verifyEmail.secondary.already-verified',
-                      uid: uid,
-                      code: request.payload.code
-                    })
-                    return P.resolve()
-                  }
-
-                  return db.verifyEmail(account, code)
-                    .then(() => {
-                      log.info({
-                        op: 'account.verifyEmail.secondary.confirmed',
-                        uid: uid,
-                        code: request.payload.code
-                      })
-                      return mailer.sendPostVerifySecondaryEmail(
-                        [],
-                        account,
-                        {
-                          acceptLanguage: request.app.acceptLanguage,
-                          secondaryEmail: matchedEmail.email
-                        })
-                    })
-                })
-            } else {
-              const isAccountVerification = butil.buffersAreEqual(code, account.emailCode)
-              let device
-
-              return db.deviceFromTokenVerificationId(uid, code)
-                .then(function (associatedDevice) {
-                  device = associatedDevice
-                }, function (err) {
-                  if (err.errno !== error.ERRNO.DEVICE_UNKNOWN) {
-                    log.error({
-                      op: 'Account.RecoveryEmailVerify',
-                      err: err,
-                      uid: uid,
-                      code: code
-                    })
-                  }
-                })
-                .then(function () {
-                  /**
-                   * Logic for account and token verification
-                   *
-                   * 1) Attempt to use code as tokenVerificationId to verify session.
-                   *
-                   * 2) An error is thrown if tokenVerificationId does not exist (check to see if email
-                   *    verification code) or the tokenVerificationId does not correlate to the
-                   *    account uid (damaged linked/spoofed account)
-                   *
-                   * 3) Verify account email if not already verified.
-                   */
-                  return db.verifyTokens(code, account)
-                    .then(function () {
-                      if (! isAccountVerification) {
-                        // Don't log sign-in confirmation success for the account verification case
-                        log.info({
-                          op: 'account.signin.confirm.success',
-                          uid: uid,
-                          code: request.payload.code
-                        })
-                        request.emitMetricsEvent('account.confirmed', {
-                          uid: uid
-                        })
-                        push.notifyUpdate(uid, 'accountConfirm')
-                      }
-                    })
-                    .catch(function (err) {
-                      if (err.errno === error.ERRNO.INVALID_VERIFICATION_CODE && isAccountVerification) {
-                        // The code is just for the account, not for any sessions
-                        return true
-                      }
-                      log.error({
-                        op: 'account.signin.confirm.invalid',
-                        uid: uid,
-                        code: request.payload.code,
-                        error: err
-                      })
-                      throw err
-                    })
-                    .then(function () {
-                      if (device) {
-                        push.notifyDeviceConnected(uid, device.name, device.id)
-                      }
-                    })
-                    .then(function () {
-
-                      // If the account is already verified, the link may have been
-                      // for sign-in confirmation or they may have been clicking a
-                      // stale link. Silently succeed.
-                      if (account.emailVerified) {
-                        return true
-                      }
-
-                      // Any matching code verifies the account
-                      return db.verifyEmail(account, account.emailCode)
-                        .then(function () {
-                          return P.all([
-                            log.notifyAttachedServices('verified', request, {
-                              email: account.email,
-                              uid: account.uid,
-                              locale: account.locale,
-                              marketingOptIn: marketingOptIn ? true : undefined
-                            }),
-                            request.emitMetricsEvent('account.verified', {
-                              uid: uid
-                            })
-                          ])
-                        })
-                        .then(function () {
-                          if (reminder === 'first' || reminder === 'second') {
-                            // if verified using a known reminder
-                            var reminderOp = 'account.verified_reminder.' + reminder
-
-                            // log to the mailer namespace that account was verified via a reminder
-                            log.info({
-                              op: 'mailer.send',
-                              name: reminderOp
-                            })
-                            return request.emitMetricsEvent('account.reminder', {
-                              uid: uid
-                            })
-                          }
-                        })
-                        .then(function () {
-                          // send a push notification to all devices that the account changed
-                          push.notifyUpdate(uid, 'accountVerify')
-                          // remove verification reminders
-                          verificationReminder.delete({
-                            uid: uid
-                          }).catch(function (err) {
-                            log.error({op: 'Account.RecoveryEmailVerify', err: err})
-                          })
-                        })
-                        .then(function () {
-                          // Our post-verification email is very specific to sync,
-                          // so only send it if we're sure this is for sync.
-                          if (service === 'sync') {
-                            return mailer.sendPostVerifyEmail(
-                              [],
-                              account,
-                              {
-                                acceptLanguage: request.app.acceptLanguage
-                              }
-                            )
-                          }
-                        })
-                    })
-                })
-            }
-          })
-          .then(
-            function () {
-              reply({})
-            },
-            reply
-          )
-      }
-    },
-    {
-      method: 'GET',
-      path: '/recovery_emails',
-      config: {
-        auth: {
-          strategy: 'sessionToken'
-        },
-        response: {
-          schema: isA.array().items(
-            isA.object({
-              verified: isA.boolean().required(),
-              isPrimary: isA.boolean().required(),
-              email: validators.email().required()
-            }))
-        }
-      },
-      handler: function (request, reply) {
-        const sessionToken = request.auth.credentials
-        const uid = sessionToken.uid
-
-        log.begin('Account.RecoveryEmailEmails', request)
-
-        if (! features.isSecondaryEmailEnabled(sessionToken.email)) {
-          return reply(error.featureNotEnabled())
-        }
-
-        db.accountEmails(uid)
-          .then(createResponse)
-          .done(reply, reply)
-
-        function createResponse(emails) {
-          return emails.map((email) => {
-            return {
-              email: email.email,
-              isPrimary: !! email.isPrimary,
-              verified: !! email.isVerified
-            }
-          })
-        }
-      }
-    },
-    {
-      method: 'POST',
-      path: '/recovery_email',
-      config: {
-        auth: {
-          strategy: 'sessionTokenWithVerificationStatus'
-        },
-        validate: {
-          payload: {
-            email: validators.email().required()
-          }
-        },
-        response: {}
-      },
-      handler: function (request, reply) {
-        const sessionToken = request.auth.credentials
-        const uid = sessionToken.uid
-        const primaryEmail = sessionToken.email
-        const ip = request.app.clientAddress
-        const email = request.payload.email
-        const emailData = {
-          email: email,
-          normalizedEmail: email.toLowerCase(),
-          isVerified: false,
-          isPrimary: false,
-          uid: uid
-        }
-
-        log.begin('Account.RecoveryEmailCreate', request)
-
-        if (! features.isSecondaryEmailEnabled(primaryEmail)) {
-          return reply(error.featureNotEnabled())
-        }
-
-        if (! sessionToken.emailVerified) {
-          return reply(error.unverifiedAccount())
-        }
-
-        if (sessionToken.tokenVerificationId) {
-          return reply(error.unverifiedSession())
-        }
-
-        if (sessionToken.email.toLowerCase() === email.toLowerCase()) {
-          return reply(error.yourPrimaryEmailExists())
-        }
-
-        customs.check(request, primaryEmail, 'createEmail')
-          .then(deleteAccountIfUnverified)
-          .then(deleteSecondaryEmailIfUnverified)
-          .then(generateRandomValues)
-          .then(createEmail)
-          .then(sendEmailVerification)
-          .then(
-            function () {
-              reply({})
-            },
-            reply
-          )
-
-        function deleteAccountIfUnverified() {
-          return db.emailRecord(email)
-            .then((emailRecord) => {
-              if (emailRecord.emailVerified) {
-                throw error.verifiedPrimaryEmailAlreadyExists()
-              }
-
-              // Check to see if this account has been unverified for longer than
-              // a day, if so, delete account so another user can add it as a
-              // secondary email
-              const msSinceCreated = Date.now() - emailRecord.createdAt
-              const minUnverifiedAccountTime = config.secondaryEmail.minUnverifiedAccountTime
-              if (msSinceCreated >= minUnverifiedAccountTime) {
-                return db.deleteAccount(emailRecord)
-              } else {
-                throw error.unverifiedPrimaryEmailNewlyCreated()
-              }
-            })
-            .catch((err) => {
-              // Email does not exist in primary account table, carry on
-              if (err.errno !== error.ERRNO.ACCOUNT_UNKNOWN) {
-                throw err
-              }
-            })
-        }
-
-        function deleteSecondaryEmailIfUnverified() {
-          return db.getSecondaryEmail(email)
-            .then((secondaryEmailRecord) => {
-              // Only delete secondary email if it is unverified and does not belong
-              // to the current user.
-              if (! secondaryEmailRecord.isVerified && ! butil.buffersAreEqual(secondaryEmailRecord.uid, uid)) {
-                return db.deleteEmail(secondaryEmailRecord.uid, secondaryEmailRecord.email)
-              }
-            })
-            .catch((err) => {
-              if (err.errno !== error.ERRNO.SECONDARY_EMAIL_UNKNOWN) {
-                throw err
-              }
-            })
-        }
-
-        function generateRandomValues() {
-          return random.hex(16)
-            .then(hex => {
-              emailData.emailCode = hex
-            })
-        }
-
-        function createEmail() {
-          return db.createEmail(uid, emailData)
-        }
-
-        function sendEmailVerification() {
-          return getGeoData(ip)
-            .then((geoData) => {
-              return mailer.sendVerifySecondaryEmail([emailData], sessionToken, {
-                code: emailData.emailCode,
-                acceptLanguage: request.app.acceptLanguage,
-                email: emailData.email,
-                primaryEmail: primaryEmail,
-                ip: ip,
-                location: geoData.location,
-                timeZone: geoData.timeZone,
-                uaBrowser: sessionToken.uaBrowser,
-                uaBrowserVersion: sessionToken.uaBrowserVersion,
-                uaOS: sessionToken.uaOS,
-                uaOSVersion: sessionToken.uaOSVersion,
-              })
-            })
-        }
-      }
-    },
-    {
-      method: 'POST',
-      path: '/recovery_email/destroy',
-      config: {
-        auth: {
-          strategy: 'sessionTokenWithVerificationStatus'
-        },
-        validate: {
-          payload: {
-            email: validators.email().required()
-          }
-        },
-        response: {}
-      },
-      handler: function (request, reply) {
-        const sessionToken = request.auth.credentials
-        const uid = sessionToken.uid
-        const primaryEmail = sessionToken.email
-        const email = request.payload.email
-
-        log.begin('Account.RecoveryEmailDestroy', request)
-
-        if (! features.isSecondaryEmailEnabled(primaryEmail)) {
-          return reply(error.featureNotEnabled())
-        }
-
-        if (sessionToken.tokenVerificationId) {
-          return reply(error.unverifiedSession())
-        }
-
-        customs.check(request, primaryEmail, 'deleteEmail')
-          .then(deleteEmail)
-          .done(() => {
-            reply({})
-          }, reply)
-
-        function deleteEmail() {
-          return db.deleteEmail(uid, email.toLowerCase())
-        }
       }
     },
     {
