+++ test/local/ip_profiling.js
@@ -54,7 +54,6 @@ var makeRoutes = function (options, requireMocks) {
   return proxyquire('../../lib/routes/account', requireMocks || {})(
     log,
     db,
-    mocks.mockBounces(),
     options.mailer || {},
     Password,
     config,
