+++ test/local/routes/account.js
@@ -50,7 +50,6 @@ var makeRoutes = function (options, requireMocks) {
   return proxyquire('../../../lib/routes/account', requireMocks || {})(
     log,
     db,
-    mocks.mockBounces(),
     options.mailer || {},
     Password,
     config,
